==> Jupyter files/2.data_preprocessing_house.ipynb <==
import numpy as np
import pandas as pd

pd.set_option("display.max_columns",None)
pd.set_option("display.max_rows",None)

df = pd.read_csv("../Dataset/houses.csv")

# shape
df.shape

df.info()

df.head()

# duplicate data
df.duplicated().sum()

df = df.drop_duplicates()

# drop unnecessary cols
df.drop(columns=['link','description','property_id','rating'],inplace=True)

df.head()

df['society'].value_counts()

df['society'].value_counts().shape

# remove rating, ★, converting to lower
import re

df["society"] = (
    df["society"]
    .apply(lambda name: re.sub(r"\d+(\.\d+)?\s?★", "", str(name)).strip())
    .str.lower()
)

df["society"].value_counts().shape

df["society"] = df["society"].str.replace("nan", "independent")

df.head()

df['price'].value_counts()

df[df['price'] == "Price on Request"]

# removed all the rows where price was not there
df = df[df["price"] != "Price on Request"]

df.head()

def treat_price(x):
    if type(x) == float:
        return x
    else:
        if x[1] == "Lac":
            return round(float(x[0]) / 100, 2)
        else:
            return round(float(x[0]), 2)

df["price"] = df["price"].str.split(" ").apply(treat_price)

df.head()

df.rename(columns={"rate": "price_per_sqft"},inplace=True)

df.head()

df["price_per_sqft"].value_counts()

df["price_per_sqft"] = df["price_per_sqft"].str.replace("₹","").str.split("/").str.get(0).str.replace(",","").str.strip().astype("float")

df.head()

df["bedRoom"].value_counts()

df['bedRoom'].isnull().sum()

df[df["bedRoom"].isnull()]

df = df[~df["bedRoom"].isnull()]

df["bedRoom"] = df["bedRoom"].str.split(" ").str.get(0).astype("int")

df.head()

df['bathroom'].value_counts()

df['bathroom'].isnull().sum()

df['bathroom'] = df['bathroom'].str.split(" ").str.get(0).astype("int")

df.head()

df['balcony'].value_counts()

df['balcony'].isnull().sum()

df['balcony'] = df['balcony'].str.split(" ").str.get(0).str.replace("No","0")

df.head()

df['additionalRoom'].value_counts()

df['additionalRoom'].isnull().sum()

df['additionalRoom'].fillna("Not Available",inplace=True)

df["additionalRoom"] = df["additionalRoom"].str.lower()

df.head()

df["noOfFloor"].value_counts()

df["noOfFloor"].isnull().sum()

df[df["noOfFloor"].isnull()]

df["noOfFloor"] = (
    df["noOfFloor"]
    .str.split(" ")
    .str.get(0)
    .replace("Ground", "0")
    .str.replace("Basement", "-1")
    .str.replace("Lower", "0")
    .str.extract(r"(\d+)")
)

df.rename(columns={"noOfFloor": "floorNum"}, inplace=True)

df.head()

df["facing"].value_counts()

df["facing"].isnull().sum()

df["facing"].fillna("N/A",inplace=True)

df.head()

## Feature Engineering

df['area'] = round((df["price"] * 10000000) / df["price_per_sqft"])

df.insert(loc=1, column="property_type", value="house")

df.head()

df.info()

df.shape

df.to_csv("../Dataset/house_cleaned.csv",index=False)
