==> Jupyter files/6.outlier_treatment.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option('display.max_columns', None)
sns.set_palette("husl")

df = pd.read_csv("../Dataset/gurgaon_properties_cleaned_v2.csv").drop_duplicates()

df.head()

## There are 2 ways to handle outliers(numerical col).
1. Normally distributed columns. (mean +- 3*std)
2. It is not normally distributed. (box plot)

df.shape

## Price

# outliers on the basis of price column
sns.distplot(df['price'])

sns.boxplot(x=df['price'])

Q1 = df['price'].quantile(0.25)
Q3 = df['price'].quantile(0.75)
IQR = Q3 - Q1

# Define bounds for outliers
lower_bound = Q1 - 1.5 * IQR
upper_bound = Q3 + 1.5 * IQR

# Identify outliers
outliers = df[(df['price'] < lower_bound) | (df['price'] > upper_bound)]

# Displaying the number of outliers and some statistics
num_outliers = outliers.shape[0]
outliers_price_stats = outliers['price'].describe()

num_outliers, outliers_price_stats

outliers.sort_values('price',ascending=False).head(20)

## on the basis of price col we can say that there are some genuine outliers but there are some data erros as well

## price_per_sq

sns.distplot(df['price_per_sqft'])

sns.boxplot(x=df['price_per_sqft'])

Q1 = df['price_per_sqft'].quantile(0.25)
Q3 = df['price_per_sqft'].quantile(0.75)
IQR = Q3 - Q1

# Define bounds for outliers
lower_bound = Q1 - 1.5 * IQR
upper_bound = Q3 + 1.5 * IQR

# Identify outliers
outliers_sqft = df[(df['price_per_sqft'] < lower_bound) | (df['price_per_sqft'] > upper_bound)]

# Displaying the number of outliers and some statistics
num_outliers = outliers_sqft.shape[0]
outliers_sqft_stats = outliers_sqft['price_per_sqft'].describe()

num_outliers, outliers_sqft_stats

outliers_sqft.head()

## there is some calculation error the area < 1000 is not in sq.ft. it is in sq.yard 

outliers_sqft['area'] = outliers_sqft['area'].apply(lambda x:x*9 if x<1000 else x)

outliers_sqft['price_per_sqft'] = round((outliers_sqft['price']*10000000)/outliers_sqft['area'])

outliers_sqft['price_per_sqft'].describe()

df.update(outliers_sqft)

sns.distplot(df['price_per_sqft'])

sns.boxplot(x=df['price_per_sqft'])

df[df['price_per_sqft']>50000]

## droping all these rows, since it is too small

df = df[df['price_per_sqft'] <= 50000]

sns.boxplot(x=df['price_per_sqft'])

## area

sns.distplot(df['area'])

sns.boxplot(x=df['area'])

df['area'].describe()

Q1 = df['area'].quantile(0.25)
Q3 = df['area'].quantile(0.75)
IQR = Q3 - Q1

# Define bounds for outliers
lower_bound = Q1 - 1.5 * IQR
upper_bound = Q3 + 1.5 * IQR

# Identify outliers
outliers_area = df[(df['area'] < lower_bound) | (df['area'] > upper_bound)]

# Displaying the number of outliers and some statistics
num_outliers = outliers_area.shape[0]
outliers_area_stats = outliers_sqft['area'].describe()

num_outliers, outliers_area_stats

df[df['area'] > 100000]

df = df[df['area'] < 100000]

sns.distplot(df['area'])

df[df['area'] > 10000].sort_values('area',ascending=False)

df.drop(index=[818, 1796, 1123, 2, 2356, 115, 3649, 2503, 1471], inplace=True)

df.loc[48,'area'] = 115*9
df.loc[300,'area'] = 7250
df.loc[2666,'area'] = 5800
df.loc[1358,'area'] = 2660
df.loc[3195,'area'] = 2850
df.loc[2131,'area'] = 1812
df.loc[3088,'area'] = 2160
df.loc[3444,'area'] = 1175

sns.distplot(df['area'])

sns.boxplot(x=df['area'])

df['area'].describe()

## bedroom

sns.distplot(df['bedRoom'])

sns.boxplot(x=df['bedRoom'])

df = df[df['bedRoom'] <= 10]

df.shape

sns.distplot(df['bedRoom'])

sns.boxplot(x=df['bedRoom'])

df['bedRoom'].describe()

## Bathroom

sns.distplot(df['bathroom'])

sns.boxplot(x=df['bathroom'])

df[df['bathroom'] > 10].sort_values('bathroom',ascending=False)

## super built up area


sns.distplot(df['super_built_up_area'])

sns.boxplot(x=df['super_built_up_area'])

df['super_built_up_area'].describe()

df[df['super_built_up_area'] > 6000]

## built up area


sns.distplot(df['built_up_area'])

sns.boxplot(x=df['built_up_area'])

df[df['built_up_area'] > 10000]

## carpet area

sns.distplot(df['carpet_area'])

sns.boxplot(x=df['carpet_area'])

df[df['carpet_area'] > 10000]

sns.distplot(df['luxury_score'])

sns.boxplot(x = df['luxury_score'])

## modify price_per_sqft with updated area value

df['price_per_sqft'] = round((df['price']*10000000)/df['area'])

sns.distplot(df['price_per_sqft'])

sns.boxplot(x = df['price_per_sqft'])

df[df['price_per_sqft'] > 42000]

x = df[df['price_per_sqft'] <= 20000]
# (x['area']/x['bedRoom']).quantile(0.02)
(x['area']/x['bedRoom']).quantile(0.05)

# Strange behaviour `area` (it is in lakh but should be in thousand) and `price_per_sqft`(same here) are giving wrong data and `carpet area` is at high range

df[df["area"]/df["bedRoom"]<250].sample(5)

sns.lmplot(data=df,x="area",y="bedRoom")

df['area_room_ratio'] = df["area"]/df["bedRoom"]

(df[df['area_room_ratio']<250])['bedRoom'].value_counts()

df[df['area_room_ratio']<100].shape

df = df[df['area_room_ratio']>100]

outlier_df = df[(df['area_room_ratio']<250) & (df['bedRoom'] > 3)]

outlier_df['bedRoom'] = round(outlier_df['bedRoom']/outlier_df['floorNum'])

df.update(outlier_df)

df['area_room_ratio'] = df["area"]/df["bedRoom"]

df[(df['area_room_ratio']<250) & (df['bedRoom'] > 4)].head()

df[(df['area_room_ratio']<250) & (df['bedRoom'] > 4)].shape

df = df[~((df['area_room_ratio']<250) & (df['bedRoom'] > 4))]

df.shape

df.head()

df.to_csv("../Dataset/gurgaon_properties_outlier_treated.csv")
