==> Jupyter files/7.missing_value_imputation.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option("display.max_columns",None)

df = pd.read_csv("../Dataset/gurgaon_properties_outlier_treated.csv")

df.head()

df.drop(columns=["Unnamed: 0"], inplace=True)

df.isnull().sum()

Observation
* built_up_area has highest missing value, so we will impute that column first.
* Approach is check super built up area and carpet area and find its ration and fill it with that ratio. so we will only keep built up area 

## Built up area

sns.scatterplot(x = df['built_up_area'],y = df['super_built_up_area'])

sns.scatterplot(x = df['built_up_area'],y = df['carpet_area'])

carpet_to_built_up_ratio = round((df['carpet_area'].dropna()/df['built_up_area'].dropna()).median(),1)
print(carpet_to_built_up_ratio)

super_to_built_up_ratio = round((df['super_built_up_area'].dropna()/df['built_up_area'].dropna()).median(),2)
print(super_to_built_up_ratio)

# only built up area is null
built_null_df = df[((~df['super_built_up_area'].isnull()) & (df['built_up_area'].isnull()) & (~df['carpet_area'].isnull()))]

built_null_df.head()

built_null_df['built_up_area'] = round(
    ((built_null_df['super_built_up_area']/super_to_built_up_ratio) + (built_null_df['carpet_area']/carpet_to_built_up_ratio))/2,2)

built_null_df.head()

df.update(built_null_df)

df.isnull().sum()

# both built up area and carpet area is null and super built up is not null
b_c_null_df = df[~(df['super_built_up_area'].isnull()) & (df['built_up_area'].isnull()) & (df['carpet_area'].isnull())]

b_c_null_df.head()

b_c_null_df['built_up_area'] = round(b_c_null_df['super_built_up_area']/super_to_built_up_ratio,2)

b_c_null_df.head()

df.update(b_c_null_df)

df.isnull().sum()

# both super built up and built up is null and carpet is not null
s_b_null_df = df[(df['super_built_up_area'].isnull()) & (df['built_up_area'].isnull()) & (~df['carpet_area'].isnull())]

s_b_null_df.head()

s_b_null_df['built_up_area'] = round((s_b_null_df['carpet_area']/carpet_to_built_up_ratio),2)

s_b_null_df.head()

df.update(s_b_null_df)

df.isnull().sum()

sns.scatterplot(x = df['built_up_area'],y = df['price'])

Observation
* All the filling has been done and the built up area to price ration should be linear but no idea why i am getting this

anamoly_df = df[(df['built_up_area'] < 2000) & (df['price'] > 2.5)][['price','area','built_up_area']]
anamoly_df.sample(5)

anamoly_df['built_up_area'] = anamoly_df['area']

df.update(anamoly_df)

sns.scatterplot(x = df['built_up_area'],y = df['price'])

Observation
* Filled the built up area 

df.drop(columns=['area','areaWithType','super_built_up_area','carpet_area','area_room_ratio'],inplace=True)

df.head()

df.isnull().sum()

## facing

1009/df.shape[0]

df['facing'].value_counts().plot(kind='pie',autopct='%0.2f%%')

df.drop(columns=['facing'],inplace=True)

df.isnull().sum()

Observation
* Facing has lot of missing value.
* from pie chart i am not get how to fill the null value there might be a relation but don't know.
* finally i am droping that column


## floornum

df[df['floorNum'].isnull()]

df[df['property_type'] == 'house']['floorNum'].median()

df['floorNum'].fillna(2.0,inplace=True)

df.isnull().sum()

Observation
* in floor num all are hosue only 2 are flat so replacing the null value with the median of house

## society

df[df['society'].isnull()]

df.drop(index=[2439],inplace=True)

## agePossession

Observation
* age posession has lot of undefined value and need to fill it
* we will go to secto and proerty type and check what are matching with undefined and will fill it with mode value of those matched result.
* By checking both still there was undefined because the combination of sector and proeprty type gives null so we pick solo columns and filled the null values.

df['agePossession'].value_counts()

df[df['agePossession'] == 'Undefined']

def mode_based_imputation(row):
    if row['agePossession'] == 'Undefined':
        mode_value = df[(df['sector'] == row['sector']) & (df['property_type'] == row['property_type'])]['agePossession'].mode()
        # If mode_value is empty (no mode found), return NaN, otherwise return the mode
        if not mode_value.empty:
            return mode_value.iloc[0] 
        else:
            return np.nan
    else:
        return row['agePossession']

df['agePossession'] = df.apply(mode_based_imputation,axis=1)

df['agePossession'].value_counts()

def mode_based_imputation2(row):
    if row['agePossession'] == 'Undefined':
        mode_value = df[(df['sector'] == row['sector'])]['agePossession'].mode()
        # If mode_value is empty (no mode found), return NaN, otherwise return the mode
        if not mode_value.empty:
            return mode_value.iloc[0] 
        else:
            return np.nan
    else:
        return row['agePossession']

df['agePossession'] = df.apply(mode_based_imputation2,axis=1)

df['agePossession'].value_counts()

def mode_based_imputation3(row):
    if row['agePossession'] == 'Undefined':
        mode_value = df[(df['property_type'] == row['property_type'])]['agePossession'].mode()
        # If mode_value is empty (no mode found), return NaN, otherwise return the mode
        if not mode_value.empty:
            return mode_value.iloc[0] 
        else:
            return np.nan
    else:
        return row['agePossession']

df['agePossession'] = df.apply(mode_based_imputation3,axis=1)


df['agePossession'].value_counts()

df.shape

df.to_csv('../Dataset/gurgaon_properties_missing_value_imputation.csv',index=False)
