==> Jupyter files/9.baseline_model.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('../Dataset/gurgaon_properties_post_feature_selection.csv')

df

Approach
* first we will do One Hot Encoding since we are going use linear model and linear model calculate distance so we will standarize the column.
* Price column is right skewed so we will use log thransformation on it

# one hot encode -> sector, balcony, agePossession, furnishing type, luxury category, floor category

X = df.drop(columns=['price'])
y = df['price']

from sklearn.model_selection import KFold, cross_val_score
from sklearn.linear_model import LinearRegression
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.svm import SVR

columns_to_encode = ['sector', 'balcony', 'agePossession', 'furnishing_type', 'luxury_category', 'floor_category']


# Applying the log1p transformation to the target variable
y_transformed = np.log1p(y)

# Creating a column transformer for preprocessing
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), ['property_type', 'bedRoom', 'bathroom', 'built_up_area', 'servant room', 'store room']),
        ('cat', OneHotEncoder(handle_unknown='ignore',drop='first'), columns_to_encode)
    ], 
    remainder='passthrough'
)

# Creating a pipeline
pipeline = Pipeline([
    ('preprocessor', preprocessor),
    ('regressor', LinearRegression())
])

# K-fold cross-validation
kfold = KFold(n_splits=10, shuffle=True, random_state=42)
scores = cross_val_score(pipeline, X, y_transformed, cv=kfold, scoring='r2')

scores.mean()

scores.std()

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y_transformed,test_size=0.2,random_state=42)

pipeline.fit(X_train,y_train)

y_pred = pipeline.predict(X_test)

y_pred = np.expm1(y_pred)

from sklearn.metrics import mean_absolute_error
mean_absolute_error(np.expm1(y_test),y_pred)

## 2nd pipeline

# Creating a pipeline
pipeline2 = Pipeline([
    ('preprocessor', preprocessor),
    ('regressor', SVR(kernel='rbf'))
])

# K-fold cross-validation
kfold = KFold(n_splits=10, shuffle=True, random_state=42)
scores = cross_val_score(pipeline2, X, y_transformed, cv=kfold, scoring='r2')

scores.mean()

scores.std()

X_train, X_test, y_train, y_test = train_test_split(X,y_transformed,test_size=0.2,random_state=42)
pipeline2.fit(X_train,y_train)

y_pred = pipeline2.predict(X_test)
y_pred = np.expm1(y_pred)
mean_absolute_error(np.expm1(y_test),y_pred)
